Ensure that the checkstyle tools are strictly an optional dependency of rewrite-java
