+++ rewrite-java/build.gradle.kts
@@ -11,6 +11,24 @@ tasks.register<JavaExec>("generateAntlrSources") {
     classpath = sourceSets["main"].runtimeClasspath
 }
 
+// So as to ensure that com.puppycrawl.tools:checkstyle is truly an optional dependency, keep things that depend on it in their own sourceSet
+sourceSets {
+    create("checkstyle") {
+        compileClasspath += sourceSets.getByName("main").output
+        runtimeClasspath += sourceSets.getByName("main").output
+    }
+    named("test") {
+        compileClasspath += sourceSets.getByName("checkstyle").output
+        runtimeClasspath += sourceSets.getByName("checkstyle").output
+    }
+}
+
+java {
+    registerFeature("checkstyle") {
+        usingSourceSet(sourceSets.getByName("checkstyle"))
+    }
+}
+
 dependencies {
     api(project(":rewrite-core"))
 
@@ -18,7 +36,8 @@ dependencies {
     api("org.jetbrains:annotations:latest.release")
 
     implementation("org.antlr:antlr4:4.8-1")
-    implementation("com.puppycrawl.tools:checkstyle:latest.release")
+    "checkstyleImplementation"("com.puppycrawl.tools:checkstyle:latest.release")
+    "checkstyleImplementation"(project(":rewrite-core"))
     implementation("commons-lang:commons-lang:latest.release")
 
     api("com.fasterxml.jackson.core:jackson-annotations:2.12.+")
@@ -27,6 +46,11 @@ dependencies {
     implementation("org.ow2.asm:asm-util:latest.release")
 
     testImplementation("org.yaml:snakeyaml:latest.release")
+    testImplementation("com.puppycrawl.tools:checkstyle:latest.release")
+}
+
+tasks.named<Jar>("jar") {
+    from(sourceSets.getByName("checkstyle").output)
 }
 
 tasks.withType<Javadoc> {
